Refs #302. Cleaned up view module.
