+++ pyrs/interface/pyrs_main.py
@@ -80,7 +80,7 @@ class PyRSLauncher(QMainWindow):
         self.ui = load_ui('pyrsmain.ui', baseinstance=self)
 
         # define
-        self.ui.pushButton_manualReduction.clicked.connect(self.do_reduce_manually)
+        self.ui.pushButton_manualReduction.clicked.connect(self.do_launch_manual_reduction)
         self.ui.pushButton_fitPeaks.clicked.connect(self.do_launch_fit_peak_window)
         self.ui.pushButton_launchTextureAnalysis.clicked.connect(self.do_launch_texture_window)
         self.ui.pushButton_launchStrainStressCalculation.clicked.connect(self.do_launch_strain_stress_window)
@@ -187,17 +187,14 @@ class PyRSLauncher(QMainWindow):
 
         return
 
-    def do_reduce_manually(self):
+    def do_launch_manual_reduction(self):
         """
         launch manual data reduction window
         :return:
         """
-        # core
-        reduction_core = pyrscore.PyRsCore()
-
         if self.manual_reduction_window is None:
             self.manual_reduction_window = manualreductionwindow.ManualReductionWindow(self)
-            self.manual_reduction_window.setup_window(reduction_core)
+            self.manual_reduction_window.setup_window()
 
         # show
         self.manual_reduction_window.show()
